Message private

+++ src/fr/uge/chatFusion/Utils/Connexion.java
@@ -1,4 +1,4 @@
 package fr.uge.chatFusion.Utils;
 
-public record Connexion(int opCode, String login) {
+public record Connexion(byte opCode, String login) {
 }
